--- test/plugin/eventwriter/event_writer_util.h
@@ -1,6 +1,0 @@
-#include <bpf/libbpf.h>
-#include <bpf/bpf.h>
-#include <windows.h>
-#include <vector>
-
-int pin_map(const char* pin_path, bpf_map* map);
