+++ test/plugin/eventwriter/packages.config
@@ -1,0 +1,10 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="eBPF-for-Windows" version="0.19.2" targetFramework="native" />
+  <package id="eBPF-for-Windows.ARM64" version="0.20.0" targetFramework="native" />
+  <package id="eBPF-for-Windows.x64" version="0.20.0" targetFramework="native" />
+  <package id="Microsoft.Windows.SDK.CPP" version="10.0.26100.2454" targetFramework="native" />
+  <package id="Microsoft.Windows.SDK.CPP.x64" version="10.0.26100.2454" targetFramework="native" />
+  <package id="Microsoft.Windows.WDK.x64" version="10.0.26100.2454" targetFramework="native" />
+  <package id="XDP-for-Windows" version="1.1.0" targetFramework="native" />
+</packages>
