--- test/plugin/eventwriter/placeholder.go
@@ -1,3 +1,0 @@
-package cprog //nolint:all
-
-// This file is a placeholder to make Go include this directory when vendoring.
